Refactor eval tests: extract shared logic to eval-runner.ts
- Create eval-runner.ts with:
  - runEvalTest(): executes test, evaluates, attaches results, records metrics
  - registerEvalTests(): registers tests for all models and test cases

- Simplify test files from ~40 lines to ~15 lines each:
  - Each file owns its describe/beforeAll/afterAll (for correct report paths)
  - Shared loop logic moved to registerEvalTests()

- Delete test-template.ts (no longer needed)

Report paths now show category folder instead of _helpers/test-template.
