+++ test/e2e/tests/assistant-eval/_helpers/eval-runner.ts
@@ -3,7 +3,7 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { test as baseTest, expect, tags } from '../../_test.setup';
+import { test as baseTest, expect } from '../../_test.setup';
 import { EvalTestCase } from '../types';
 import { evaluateWithLLM } from './llm-grader';
 import { formatResultsHtml } from './format-results';
@@ -12,77 +12,10 @@ import { recordAssistantEval, parseToolsFromResponse, type AssistantEvalInput } 
 import { type Application, type MultiLogger, type Sessions, type HotKeys, type TestTeardown } from '../../../infra';
 import { type Settings } from '../../../fixtures/test-setup/settings.fixtures';
 
-// Re-export for test files
-export { tags };
-
 /**
- * Options for configuring eval tests.
+ * Fixtures required for running eval tests.
  */
-export interface EvalTestsOptions {
-	/** Category for metrics (e.g., 'notebooks', 'tools', 'hallucination'). Defaults to 'general'. */
-	category?: string;
-}
-
-/**
- * Sets up the eval test structure inside a test.describe block.
- * The test.describe must be in the test file for UI visibility.
- *
- * @example
- * import { test } from '../../_test.setup';
- * import { tags, evalTests } from '../_helpers/test-template';
- * import { myTestCase } from './my-test-case';
- *
- * test.describe('Assistant Eval: Category', { tag: [tags.ASSISTANT_EVAL] }, () => {
- *   evalTests(test, [
- *     myTestCase,
- *   ], { category: 'notebooks' });
- * });
- */
-export function evalTests(
-	test: typeof baseTest,
-	testCases: EvalTestCase[],
-	options: EvalTestsOptions = {}
-): void {
-	const { category = 'general' } = options;
-	test.beforeAll(async ({ assistant }) => {
-		await assistant.openPositronAssistantChat();
-		await assistant.loginModelProvider('anthropic-api');
-	});
-
-	const modelKeys = getModelKeys();
-
-	modelKeys.forEach((modelKey, index) => {
-		const modelConfig = getModelConfig(modelKey);
-
-		test.describe(modelKey, () => {
-			test.beforeAll(async ({ assistant }) => {
-				if (index > 0 || modelKey !== 'sonnet') {
-					await assistant.selectChatModel(modelConfig.pickerName);
-				}
-			});
-
-			testCases.forEach((testCase) => {
-				test(`${testCase.id}: ${testCase.description}`,
-					{ tag: testCase.tags ?? [] },
-					async ({ app, sessions, hotKeys, cleanup, settings, logger }) => {
-						await runEvalTest(testCase, modelKey, modelConfig.displayName, category, {
-							app, sessions, hotKeys, cleanup, settings, logger
-						});
-					}
-				);
-			});
-		});
-	});
-
-	test.afterAll(async ({ assistant }) => {
-		await assistant.logoutModelProvider('anthropic-api');
-	});
-}
-
-/**
- * Runs a single eval test case and records metrics.
- */
-interface RunEvalTestFixtures {
+export interface EvalTestFixtures {
 	app: Application;
 	sessions: Sessions;
 	hotKeys: HotKeys;
@@ -91,12 +24,21 @@ interface RunEvalTestFixtures {
 	logger: MultiLogger;
 }
 
-async function runEvalTest(
+/**
+ * Runs a single eval test case: executes the test, evaluates with LLM, attaches results, records metrics.
+ *
+ * @param testCase The test case to run
+ * @param modelKey Short model key (e.g., 'sonnet', 'opus')
+ * @param modelDisplayName Display name for reports (e.g., 'claude sonnet 4')
+ * @param category Category for metrics (e.g., 'notebooks', 'tools')
+ * @param fixtures Playwright test fixtures
+ */
+export async function runEvalTest(
 	testCase: EvalTestCase,
 	modelKey: string,
 	modelDisplayName: string,
 	category: string,
-	fixtures: RunEvalTestFixtures
+	fixtures: EvalTestFixtures
 ): Promise<void> {
 	const { app, sessions, hotKeys, cleanup, settings, logger } = fixtures;
 
@@ -150,4 +92,55 @@ async function runEvalTest(
 	};
 
 	await recordAssistantEval(metricInput, responseDurationMs, !!app.code.electronApp, logger);
+}
+
+/**
+ * Registers eval tests for all configured models and test cases.
+ * Call this inside a test.describe block after setting up beforeAll/afterAll.
+ *
+ * @param test The Playwright test object
+ * @param testCases Array of test cases to run
+ * @param category Category for metrics (e.g., 'notebooks', 'tools', 'hallucination')
+ *
+ * @example
+ * test.describe('Assistant Eval: Notebooks', { tag: [...] }, () => {
+ *   test.beforeAll(async ({ assistant }) => {
+ *     await assistant.openPositronAssistantChat();
+ *     await assistant.loginModelProvider('anthropic-api');
+ *   });
+ *
+ *   registerEvalTests(test, testCases, 'notebooks');
+ *
+ *   test.afterAll(async ({ assistant }) => {
+ *     await assistant.logoutModelProvider('anthropic-api');
+ *   });
+ * });
+ */
+export function registerEvalTests(
+	test: typeof baseTest,
+	testCases: EvalTestCase[],
+	category: string
+): void {
+	getModelKeys().forEach((modelKey, index) => {
+		const modelConfig = getModelConfig(modelKey);
+
+		test.describe(modelKey, () => {
+			test.beforeAll(async ({ assistant }) => {
+				if (index > 0 || modelKey !== 'sonnet') {
+					await assistant.selectChatModel(modelConfig.pickerName);
+				}
+			});
+
+			testCases.forEach((testCase) => {
+				test(`${testCase.id}: ${testCase.description}`,
+					{ tag: testCase.tags ?? [] },
+					async ({ app, sessions, hotKeys, cleanup, settings, logger }) => {
+						await runEvalTest(testCase, modelKey, modelConfig.displayName, category, {
+							app, sessions, hotKeys, cleanup, settings, logger
+						});
+					}
+				);
+			});
+		});
+	});
 }
