+++ test/e2e/tests/assistant-eval/hallucination/eval-hallucination.test.ts
@@ -3,16 +3,27 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { test } from '../../_test.setup';
-import { evalTests, tags } from '../_helpers/test-template';
+import { test, tags } from '../../_test.setup';
+import { registerEvalTests } from '../_helpers/eval-runner';
 import { rForestedHallucination } from './r-forested-hallucination';
 import { pythonNoExecutionHallucination } from './python-no-execution-hallucination';
+
+const testCases = [
+	rForestedHallucination,
+	pythonNoExecutionHallucination,
+];
 
 test.use({ suiteId: __filename });
 
 test.describe('Assistant Eval: Hallucination', { tag: [tags.ASSISTANT_EVAL] }, () => {
-	evalTests(test, [
-		rForestedHallucination,
-		pythonNoExecutionHallucination,
-	], { category: 'hallucination' });
+	test.beforeAll(async ({ assistant }) => {
+		await assistant.openPositronAssistantChat();
+		await assistant.loginModelProvider('anthropic-api');
+	});
+
+	registerEvalTests(test, testCases, 'hallucination');
+
+	test.afterAll(async ({ assistant }) => {
+		await assistant.logoutModelProvider('anthropic-api');
+	});
 });
