+++ test/e2e/tests/assistant-eval/notebooks/eval-notebooks.test.ts
@@ -3,22 +3,35 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { test } from '../../_test.setup';
-import { evalTests, tags } from '../_helpers/test-template';
+import { test, tags } from '../../_test.setup';
+import { registerEvalTests } from '../_helpers/eval-runner';
 import { rNotebookAutomaticContext } from './r-notebook-automatic-context';
 import { rNotebookGetCells } from './r-notebook-get-cells';
 import { rNotebookEditCells } from './r-notebook-edit-cells';
 import { rNotebookRunCells } from './r-notebook-run-cells';
 import { rNotebookCreate } from './r-notebook-create';
+import { pyNotebookGetCells } from './py-notebook-get-cells';
+
+const testCases = [
+	rNotebookAutomaticContext,
+	rNotebookGetCells,
+	rNotebookEditCells,
+	rNotebookRunCells,
+	rNotebookCreate,
+	pyNotebookGetCells,
+];
 
 test.use({ suiteId: __filename });
 
 test.describe('Assistant Eval: Notebooks', { tag: [tags.ASSISTANT_EVAL, tags.POSITRON_NOTEBOOKS] }, () => {
-	evalTests(test, [
-		rNotebookAutomaticContext,
-		rNotebookGetCells,
-		rNotebookEditCells,
-		rNotebookRunCells,
-		rNotebookCreate,
-	], { category: 'notebooks' });
+	test.beforeAll(async ({ assistant }) => {
+		await assistant.openPositronAssistantChat();
+		await assistant.loginModelProvider('anthropic-api');
+	});
+
+	registerEvalTests(test, testCases, 'notebooks');
+
+	test.afterAll(async ({ assistant }) => {
+		await assistant.logoutModelProvider('anthropic-api');
+	});
 });
